requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.17.0
joblib>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0
pydub>=0.25.0
tensorflow>=2.13.0
opencv-python>=4.8.0
